rename some files in vacancies module
