+++ app/modules/vacancies/view/vacancyView.js
